Cria módulo para REPL
